backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "click>=8.3.1",
    "cohere>=5.20.1",
    "fastapi>=0.115.0",
    "httpx>=0.28.1",
    "openai>=2.14.0",
    "openai-agents>=0.6.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "uvicorn>=0.34.0",
]
